CartaStyles

Documentation

styleCartaItems :: ToHtml m => Int -> [m] -> Html ()

styleCartaItems n is produces Html using the ToHtml instances of is elements and gathering them in rows of n columns (Bootstrap or Zurb Foundation).

Note: calling styleCartaItems 0 is will do no extra wrapping around the HTML produced by each item.