CartaTypes

Description

En este modulo se describen los tipos bases para describir un elemento de una carta de restaurante. Por simplicidad se ha definido que un elemento de carta tiene nombre, posiblemente descripcion y al menos un precio. Para otro tipo de elementos con mas precios como para describir un trago en la seccion de licores se ha dispuesto dos tipos mas que pueden tener dos o tres precios.

Documentation

data ItemCarta

Un item de carta es un item descrito con 3,2 o 1 precio.

Constructors

D3PItem Item3

D2PItem Item2

D1PItem Item

data Item

Constructors

Item

Fields

itemName :: String

description :: Maybe String

firstPrice :: String

data Item2

Constructors

Item2

Fields

itemInfo :: Item

secondPrice :: String

data Item3

Constructors

Item3

Fields

item2Info :: Item2

thirdPrice :: String

class PricedOnce a where

Un item o elemento de la carta/menu que tiene al menos un primer precio

Methods

priceOne :: a -> String

class PricedTwice a where

Un item o elemento de la carta/menu que tiene al menos un primer precio

Methods

priceTwo :: a -> String

class NamedItem a where

Methods

itemsName :: a -> String

itemsDescription :: a -> Maybe String