CsvUtils

Documentation

readRecords :: FromRecord a => String -> IO [Either String a]

readRecords fname parses each record of a csv file named fname independently of wether or not parsing other records succed.; Note: the FromRecord instance should be specified e.g readRecords "people.csv" :: IO [Either String Person] where person would be a FromRecord instance.

readNamedRecords :: FromNamedRecord a => String -> IO [Either String (Header, a)]

readNamedRecords parses each every named record independently of wether or not parsing other records succed.

readNamedRecords' :: FromNamedRecord b => String -> IO [Either String b]

printRecords :: Show a1 => IO [Either a a1] -> IO ()

print Records in a separate line

printNamedRecords :: Show a2 => IO [Either a1 (a, a2)] -> IO ()