FileName

Documentation

nameAndExt :: [Char] -> ([Char], [Char])

returns a (name, ext) of file name

getName :: [Char] -> [Char]

getExt :: [Char] -> [Char]

modifyName :: ([Char] -> [Char]) -> [Char] -> [Char]

modifyName f applies f to the filename excluding the extension

modifyExt :: ([Char] -> [Char]) -> [Char] -> [Char]

modifyExt f applies f to the extension of a file name.

applySafe :: ([t] -> [t]) -> [t] -> [t]

applySafe f ls applies f to ls only if it doesnt emply an empty list as a result.

nameWithExt :: ([Char], [Char]) -> [Char]

unwordsWithCommas :: [String] -> String

nameIs :: ([Char] -> Bool) -> [Char] -> Bool

nameIsNot :: [Char] -> [Char] -> Bool

extIs :: ([Char] -> Bool) -> [Char] -> Bool

extIsNot :: [Char] -> [Char] -> Bool

isElemOf :: Eq a => [a] -> a -> Bool

isSymbolic :: Char -> Bool

nameToLower :: [Char] -> [Char]

nameToUpper :: [Char] -> [Char]

replaceIf :: (b -> Bool) -> b -> [b] -> [b]

replaceChar :: Eq b => b -> b -> [b] -> [b]

replaceSpacesBy :: Char -> [Char] -> [Char]

capitalize :: String -> String

deleteSpace :: [Char] -> [Char]

underscoreAndLower :: [Char] -> [Char]

camelCase :: [Char] -> [Char]

concatWith :: [a] -> [a] -> [a] -> [a]

cleanUpFileName :: [Char] -> [Char]

cleanUp :: [Char] -> [Char]