GenericUtils

Description

Mostly math functions and some higher order combinators that will probably be use in future proyects.

Documentation

linspace :: (Fractional a, Ord a) => a -> a -> Integer -> [a]

evalInRange :: (Fractional a, Ord a) => a -> a -> (a -> b) -> Integer -> [b]

evalInRange' :: (Double, Double) -> (Double -> b) -> Integer -> [b]

within :: Ord a => a -> a -> a -> Ordering

within i s x es una funcion que prueba si x esta en el interval definido entre un limite inferior i y uno superior s

within' :: Ord a => a -> a -> a -> Bool

withinCentered :: (Num a, Ord a) => a -> a -> a -> Ordering

withinCentered' :: (Num a, Ord a) => a -> a -> a -> Bool

modf :: (Num a, Ord a) => a -> a -> a

constrain :: Ord a => a -> a -> a -> a

constrain some times called clamp is function that forces a value to be in interval

mapFromRange :: Fractional a => (a, a) -> (a, a) -> a -> a

mapFromRange i o x positions x in interval o proportionally to its location on i.

(...) :: (b -> c) -> (a -> a1 -> b) -> a -> a1 -> c

Double composition operator apply a function that takes a single arg after a function that takes 2 args For ex: a dotProduct using regular function composition would be written as dotProduct xs = sum . zipWith (*) xs wich has type [a] -> [a] -> a with double composition dotProduct = sum ... zipWith (*)

(<..) :: (c -> r) -> (a -> b -> c) -> a -> b -> r

For more nemotecnic operators

(..>) :: (a -> s -> c) -> (c -> r) -> a -> s -> r

(<.) :: (b -> c) -> (a -> b) -> a -> c

Regular function composition

(.>) :: (a -> b) -> (b -> c) -> a -> c

Reverse composition

fmap2 :: (Functor f, Functor g) => (a -> b) -> f (g a) -> f (g b)

fmap to a nested functor i.e f over IO (Maybe a) not to be mistake with liftA2 witch has a different type signature.

toFloat :: Integral a => a -> Float

toDouble :: Integral a => a -> Double

dotProduct :: [Double] -> [Double] -> Double

rotate :: Double -> (Double, Double) -> [Double]