GetCmdOpts

Documentation

renamingOpts :: IO [Flag]

renamingOpts delivers the command line specified flags/options

data Flag

A command line flag is on of the following:

Constructors

TrimBeg Int
Trim a number of char off starting from left to right

TrimEnd Int
Trim n char from the end

ListUsing String
Enumerate files prepending a constant string

Enumerate
Enumerate files at the end

EnumerateBeg

Delete String
Delete occurrences of a substring

EnumPrepending String

EnumAppending String

Clean
Clean up file name replace white space for _ remove sybolic chars, remove trailing spaces.

Append String
Append a string to each file name

Prepend String
Prepend a string to each file name

Replace String String
Replace all instances of

FileSelection String

GroupEnum

makeReplace :: [Char] -> Flag