ListModifiers

Documentation

swap :: (t1, t) -> (t, t1)

applyToFirst :: (t -> t1) -> (t, t2) -> (t1, t2)

applyToSecond :: (t -> t2) -> (t1, t) -> (t1, t2)

applyToBoth :: (t -> t1) -> (t, t) -> (t1, t1)

concatTuple :: ([a], [a]) -> [a]

concatTupleUsing :: [a] -> ([a], [a]) -> [a]

tuplify2 :: [t] -> (t, t)

tuplify2With :: (t -> t1) -> (t -> t2) -> [t] -> (t1, t2)

tuplify3 :: [t] -> (t, t, t)

tuplify3With :: (t -> t1) -> (t -> t2) -> (t -> t3) -> [t] -> (t1, t2, t3)

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

curry3 :: ((a, b, c) -> d) -> a -> b -> c -> d

toHead :: (a -> a) -> [a] -> [a]

toLast :: (r -> r) -> [r] -> [r]

mapInit :: (a -> a) -> [a] -> [a]

splitEvery :: Int -> [t] -> [[t]]

splitEvery n creates a list with sublists of length n.

intersperseEvery :: Int -> a -> [a] -> [a]

drop in an element every n elements

insertAt :: Int -> a -> [a] -> [a]

InsertAt n c inserts element c at index n

innerZip :: [a] -> [(a, a)]

innerZip takes a list and returns a a list of adjacent elements If the input list has odd length then the last element is trimmed.

innerZip3 :: [a] -> [(a, a, a)]

modifyWord :: Eq a => ([a] -> [a]) -> [a] -> [a] -> [a]

modifyWord f w modifies every ocurrence of w with f

containsWord :: Eq a => [a] -> [a] -> Bool

replaceWord :: Eq a => [a] -> [a] -> [a] -> [a]

replaceWord w nw replaces every occurence of w with nw

replaceWord w nw = modifyWord (const nw) w

deleteWord :: Eq a => [a] -> [a] -> [a]

deleteHeadWhile :: (t -> Bool) -> [t] -> [t]

deleteLastWhile :: (t -> Bool) -> [t] -> [t]

deleteEndsWhile :: (t -> Bool) -> [t] -> [t]

deleteExtra :: Eq a => [a] -> [a] -> [a]

decimate :: [a] -> [a]

decimate collects all even indecis of a list

everyEven :: [a] -> [a]

everyOdd :: [a] -> [a]

equalElems :: Eq a => [a] -> [a] -> Int

sharedElemsPerc :: (Eq a, Fractional a1) => [a] -> [a] -> a1

areSimilarBy :: (Eq a, Fractional r, Ord r) => r -> [a] -> [a] -> Bool

groupSimilarBy :: (Fractional r, Ord a, Ord r) => r -> [[a]] -> [[[a]]]

equalLengthSubLists :: [[a]] -> [[a]]

zipEq :: [Char] -> [Char] -> [Char]

withEqLen :: [[Char]] -> [[Char]]

allOf :: [a -> Bool] -> a -> Bool

oneOf :: [a -> Bool] -> a -> Bool

combinePredicates :: ([Bool] -> Bool) -> [a -> Bool] -> a -> Bool