Main

Documentation

trimFileEnd :: Int -> String -> String

trimFileBeg :: Int -> [Char] -> [Char]

appendToFileName :: [Char] -> [Char] -> [Char]

prependToFileName :: [Char] -> [Char] -> [Char]

enumerateNames :: Bool -> Bool -> String -> [String] -> [String]

enumEnd :: [String] -> [String]

enumBeg :: [String] -> [String]

enumAndUnderscore :: [[Char]] -> [String]

enumAndUnderscore' :: [[Char]] -> [String]

enumPrepending :: [Char] -> [[Char]] -> [String]

enumAppending :: [Char] -> [[Char]] -> [String]

enumSimilarBy :: (Ord s, Fractional s) => s -> [[Char]] -> [String]

renameFiles :: (String -> Bool) -> FilePath -> ([String] -> [String]) -> IO ()

renameFiles p path f renames all files satisfying p at path with f.

renameAllFilesAt :: FilePath -> ([String] -> [String]) -> IO ()

renameAllFilesUsing :: ([String] -> [String]) -> IO ()

renameCopying :: (String -> Bool) -> String -> FilePath -> ([String] -> [String]) -> IO ()

renameAllCopyingAt :: String -> FilePath -> ([String] -> [String]) -> IO ()

renameAllCopiedFilesUsing :: ([String] -> [String]) -> IO ()

renameFilesWithExtension :: (String -> Bool) -> ([String] -> [String]) -> IO ()

performOnSelection :: FileSelector -> Flag -> IO ()

modifierFromOption :: Flag -> [String] -> [String]

onListOrElementwise :: SelectionModifier -> [String] -> [String]

selectOption :: Flag -> SelectionModifier

data SelectionModifier

SelectionModifier wraps functions that operate on each file name in isolation or functions that depend on the context of other file names

Constructors

Elementwise (String -> String)

OnList ([String] -> [String])

data FileSelector

A type to specify a kind of selection

Constructors

WithExt String

WithSubstring String

OnAll

fileSelectorFromArg :: Flag -> FileSelector

isFileSelecion :: Flag -> Bool

getFileSelection :: [Flag] -> Maybe Flag

main :: IO ()