ParsecRead

Description

This module is a utility module for defining Read instances.

The general procedure for making X an instance of Read is to first make it an instance of ParsecRead and then doing:

instance X Read where readsPrec _ = parsecToReadS

Although the default instance (instance (ParsecRead a) => Read a where readsPrec _ = parsecToReadS) could be written to derive instances automatically it involves language extensions (UndecidableInstances, FlexibleInstances) and some other trickery.

Documentation

class ParsecRead a where

Minimal complete definition

parsecRead

Methods

parsecToReadS :: String -> [(a, String)]

parsecRead :: ParserR a

readF :: (Functor f, Read a) => f String -> f a

A simple utility to read inside functors.

type ParserR a = Parsec String () a